@@ -1,9 +1,8 @@
 import { NgModule }       from '@angular/core';
 import { BrowserModule  } from '@angular/platform-browser';
-import { RouterModule } from '@angular/router';
 import { FormsModule } from '@angular/forms';
-import { PairsModule } from 'src/pairs.module';
-import { PairsComponent } from 'src/pairs.component';
+import { PairsModule } from 'src/pairs/pairs.module';
+import { PairsComponent } from 'src/pairs/pairs.component';
 
 @NgModule({
     declarations: [],
